New sprites, rendering changes, and more
Added more of Levi's sprites: his symbiote (not in use currently) some
blocks, and the gun skill icon.

Added skill icons and cooldowns.

Entities are now drawn top to bottom, so a lower entity will be drawn
over a higher entity.

There are now walls that can be collided with, stood in front of, and
(sort of) hidden behind.

Added a bunch of TODOs for bugs and things that need to be dealt with.

Maybe some other stuff? I dunno.